Split GetUpgradeContext to a dedicated file
